Подправил все запросы к итоговому виду
