+++ packages/proto/main.proto
@@ -9,7 +9,6 @@ service AuroraLauncherService {
   rpc getServers(.google.protobuf.Empty) returns (ServersResponse);
   rpc getProfile(ProfileRequest) returns (ProfileResponse);
   rpc getUpdates(UpdateRequest) returns (UpdateResponse);
-  //rpc verify(VerifyRequest) returns (VerifyResponse);
 }
 
 message AuthRequest {
@@ -22,8 +21,8 @@ message AuthResponse {
   string accessToken = 3;
   string token = 4;
   optional bool isAlex = 5;
-  optional string skinURL = 6;
-  optional string capeURL = 7;
+  optional string skinUrl = 6;
+  optional string capeUrl = 7;
 }
 message ServersResponse {
   repeated Server servers = 1;
@@ -52,7 +51,7 @@ message ProfileResponse {
   repeated string update = 14;
   repeated string updateVerify = 15;
   repeated string updateExclusions = 16;
-  WhiteListType whiteListType = 17;
+  string whiteListType = 17;
   optional int32 whiteListPermisson = 18;
   optional string whiteListUUIDs = 19;
 }
@@ -65,35 +64,18 @@ message ServerConfig {
 message ProfileLibrary {
   string path = 1;
   string sha1 = 2;
-  TypeLibrary type = 3;
-  optional LibraryRule rules = 4;
+  string type = 3;
+  repeated LibraryRule rules = 4;
   optional bool ignoreClassPath = 5;
 }
-enum TypeLibrary {
-  reserved "library", "native";
-}
 message LibraryRule {
-  Action action = 1;
+  string action = 1;
   optional OS os = 2;
 }
-enum Action {
-  allow = 0;
-  disallow = 1;
-}
 message OS {
-  OSName name = 1;
+  string name = 1;
   optional string arch = 2;
   optional string version = 3;
-}
-enum OSName {
-  linux = 0;
-  osx = 1;
-  windows = 2;
-}
-enum WhiteListType {
-  null = 0;
-  uuids = 1;
-  permissions = 2;
 }
 message UpdateRequest {
   string dir = 1;
