+++ packages/server/src/components/api/grpc/index.ts
@@ -1,0 +1,2 @@
+export * from "./GrpcServerManager";
+export * from "./Requests";
