--- packages/server/src/components/api/websocket/ExtendedWebSocketClient.ts
@@ -1,8 +1,0 @@
-import { WebSocketClient } from "aurora-rpc-server";
-
-export interface ExtendedWebSocketClient extends WebSocketClient {
-    isAuthed: boolean;
-
-    isVerified: boolean;
-    verifyToken?: string;
-}
