--- packages/server/src/components/api/websocket/index.ts
@@ -1,2 +1,0 @@
-export * from "./WebManager";
-//export * from "./requests";
