--- packages/server/src/components/api/websocket/requests/index.ts
@@ -1,5 +1,0 @@
-export * from "./AuthRequest";
-export * from "./ProfileRequest";
-export * from "./ServersRequest";
-export * from "./UpdatesRequest";
-export * from "./VerifyRequest";
